clean up the package org.hibernate.engine.transaction
- it had classes/interfaces which are already unused in H5!
- move TransactionObserver and IsolationDelegate to a much
  more sensible location next to TransactionCoordinator

This will break custom implementations of TransactionCoordinator
but I highly doubt there are very many of those!
