+++ hibernate-core/src/main/java/org/hibernate/resource/transaction/spi/IsolationDelegate.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later.
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.engine.transaction.spi;
+package org.hibernate.resource.transaction.spi;
 
 import org.hibernate.HibernateException;
 import org.hibernate.jdbc.WorkExecutorVisitable;
